relocate variable and function validator
