+++ tests/test_validator.py
@@ -33,3 +33,23 @@ class TestValidator(unittest.TestCase):
             }
         ]
         self.assertTrue(data_structure)
+
+    def test_is_variable(self):
+        var1 = 123
+        var2 = "abc"
+        self.assertTrue(validator.is_variable(("var1", var1)))
+        self.assertTrue(validator.is_variable(("var2", var2)))
+
+        __var = 123
+        self.assertFalse(validator.is_variable(("__var", __var)))
+
+        func = lambda x: x + 1
+        self.assertFalse(validator.is_variable(("func", func)))
+
+        self.assertFalse(validator.is_variable(("unittest", unittest)))
+
+    def test_is_function(self):
+        func = lambda x: x + 1
+        self.assertTrue(validator.is_function(("func", func)))
+
+        self.assertTrue(validator.is_function(("func", validator.is_testcase)))
